Unir codigo y mover excepciones
