+++ src/main/java/com/entrevistador/orquestador/infrastructure/rest/advice/ExcepcionGlobalHandler.java
@@ -1,10 +1,10 @@
 package com.entrevistador.orquestador.infrastructure.rest.advice;
 
+import com.entrevistador.orquestador.dominio.excepciones.PdfException;
 import com.entrevistador.orquestador.infrastructure.adapter.dto.ErrorResponse;
 import com.entrevistador.orquestador.dominio.model.enums.CodigoErrorEnum;
-import com.entrevistador.orquestador.infrastructure.adapter.exceptions.EntrevistaProcesoException;
-import com.entrevistador.orquestador.infrastructure.adapter.exceptions.HojaDeVidaInvalidaException;
-import com.entrevistador.orquestador.infrastructure.adapter.exceptions.LimiteEntrevistaException;
+import com.entrevistador.orquestador.dominio.excepciones.EntrevistaProcesoException;
+import com.entrevistador.orquestador.dominio.excepciones.LimiteEntrevistaException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
@@ -29,8 +29,8 @@ public class ExcepcionGlobalHandler {
         return Mono.just(new ResponseEntity<>(errorResponse, HttpStatus.BAD_REQUEST));
     }
 
-    @ExceptionHandler(HojaDeVidaInvalidaException.class)
-    public Mono<ResponseEntity<ErrorResponse>> handleHojaDeVidaInvalidaException(HojaDeVidaInvalidaException ex) {
+    @ExceptionHandler(PdfException.class)
+    public Mono<ResponseEntity<ErrorResponse>> handleHojaDeVidaInvalidaException(PdfException ex) {
         ErrorResponse errorResponse = new ErrorResponse(LocalDateTime.now(), ex.getCode(), ex.getMessage());
         return Mono.just(new ResponseEntity<>(errorResponse, HttpStatus.BAD_REQUEST));
     }
